+++ tests/integrations/test_metric_lightning.py
@@ -14,7 +14,7 @@
 import torch
 from pytorch_lightning import Trainer
 
-from tests.integrations.boring_model import BoringModel
+from tests.integrations.lightning_models import BoringModel
 from torchmetrics import Metric
